requirements.txt
absl-py==2.3.1
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anytree==2.12.1
async-timeout==5.0.1
attrs==25.3.0
blobfile==3.0.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
datasets==2.14.7
dill==0.3.7
docker-pycreds==0.4.0
docstring_parser==0.16
evaluate==0.4.1
filelock==3.13.1
frozenlist==1.6.0
fsspec==2023.10.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-auth-oauthlib==1.0.0
grpcio==1.74.0
h5py==3.13.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
lightning-utilities==0.14.3
lxml==5.4.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.15
networkx==3.3
numpy==2.2.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
packaging==25.0
pandas==2.2.3
peft==0.14.0
pip==25.1
platformdirs==4.3.7
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodomex==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.0.4
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
responses==0.18.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.13.0
sentencepiece==0.2.0
sentry-sdk==2.26.1
setproctitle==1.3.5
setuptools==78.1.1
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tensorboard==2.12.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.4.0+cu121
torchmetrics==1.7.1
tqdm==4.67.1
transformers==4.49.0
triton==3.0.0
trl==0.11.4
typeguard==4.4.2
typing_extensions==4.13.2
typing-inspection==0.4.0
tyro==0.9.19
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.9
Werkzeug==3.1.3
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.0
